Refactoring the distance cache
