+++ openquake/hazardlib/calc/filters.py
@@ -26,6 +26,8 @@ from scipy.interpolate import interp1d
 
 from openquake.baselib.python3compat import raise_
 from openquake.hazardlib import site
+from openquake.hazardlib.geo.surface.multi import (
+    _multi_distances, _multi_rx_ry0)
 from openquake.hazardlib.geo.utils import (
     KM_TO_DEGREES, angular_distance, fix_lon, get_bounding_box,
     get_longitudinal_extent, BBoxError, spherical_to_cartesian)
@@ -44,11 +46,59 @@ def magstr(mag):
     return '%.2f' % numpy.float32(mag)
 
 
-def get_distances(rupture, sites, param):
+def get_dist(rup, sites, param, dcache):
+    """
+    Calculates the distances for multi-surfaces using a cache.
+
+    :param rup:
+        An instance of :class:`openquake.hazardlib.source.rupture.BaseRupture`
+    :param sites:
+        A list of sites or a site collection
+    :param param:
+        The required rupture-distance parameter
+    :param dcache:
+        A dictionary with the distances. The first key is the
+        surface ID and the second one is the type of distance. In a traditional
+        calculation dcache is instatianted by in the `get_ctxs` method of the
+        :class:`openquake.hazardlib.contexts.ContextMaker`
+    :returns:
+        A dictionary with the computed distances for the rupture in input
+    """
+    # Update the distance cache
+    suids = []  # surface IDs
+    for srf in rup.surface.surfaces:
+        suids.append(srf.suid)
+        if srf.suid not in dcache:
+            # This function returns the distances that will be added to the
+            # cache. In case of Rx and Ry0, the information cache will
+            # include the ToR of each surface as well as the GC2 t and u
+            # coordinates for each section.
+            distdic = _multi_distances(srf, sites, param)
+            # Save information into the cache for the current surfac.
+            for key in distdic:
+                dcache[srf.suid, key] = distdic[key]
+
+    # Computing distances using the cache
+    if param in ['rjb', 'rrup']:
+        distances = dcache[suids[0], param]
+        # This is looping over all the surface IDs composing the rupture
+        for suid in suids[1:]:
+            distances = numpy.minimum(distances, dcache[suid, param])
+    elif param in ['rx', 'ry0']:
+        # The computed distances. In this case we are not going to add them to
+        # the cache since they cannot be reused
+        distances = _multi_rx_ry0(dcache, suids, param)
+    else:
+        raise ValueError("Unknown distance measure %r" % param)
+    return distances
+
+
+def get_distances(rupture, sites, param, dcache=None):
     """
     :param rupture: a rupture
     :param sites: a mesh of points or a site collection
     :param param: the kind of distance to compute (default rjb)
+    :param dcache: a dictionary (surfaceID, dist_type) -> distances
     :returns: an array of distances from the given sites
     """
     if not rupture.surface:  # PointRupture
